Begin redesiging the navigation bar
