+++ app/assets/stylesheets/pages/premium_page.scss
@@ -1,5 +1,4 @@
 .premium-tab {
-  background-color: #EA9A1A !important;
   a {
     background-color: #EA9A1A !important;
     color: #FFFFFF !important;
@@ -59,6 +58,7 @@
 li.active.premium-tab > a,
 li.premium-tab > a:hover {
   border-top: 3px solid #875A12 !important;
+  padding-top: 0 !important;
 }
 .premium-tab a.active,
 .premium-tab a:hover {
