+++ client/app/assets/styles/navbar.scss
@@ -17,16 +17,8 @@
   }
 
   li {
-    cursor: pointer;
-    &:not(:last-of-type) {
-      border-right: solid 1px #d4d4d4;
-    }
-    background-color: #e9e9e9;
     padding: 0px;
     white-space: nowrap;
-    &:hover {
-      background-color: white;
-    }
     &.active {
       a {
         margin-top: -3px;
@@ -39,12 +31,6 @@
         position: absolute;
         top: 0;
         left: 0;
-        &.premium-tab {
-          background-color: #875A12
-        }
-        &:not(.premium-tab) {
-          background-color: #027e6b;
-        }
       }
       position: relative;
       background-color: white;
